[Sigma BugID #42200] union: refine u-boot for union project
1. add union_emmc_defconfig, copying from union_veloce_defconfig
   with minimum driver enabled.
   TODO: enable necessary drivers for union
   -CONFIG_TRIX_NONE=y
   +CONFIG_BOARD_CONFIGURATION=y
   +CONFIG_SYS_BOOTMIPS=y
   +CONFIG_SIGMA_AUTOUPDATE=y
   +CONFIG_PANEL_ID_EXTENSION=y
   -# CONFIG_CMD_NET is not set
   -# CONFIG_CMD_NFS is not set
   -CONFIG_REGEX=y
   +CONFIG_NETDEVICES=y
   +CONFIG_ETH_DESIGNWARE=y
   +CONFIG_USB=y
   +CONFIG_DM_USB=y
   +CONFIG_USB_EHCI_HCD=y
   +CONFIG_USB_EHCI_TRIX=y
2. general error/warning fixes for aarch64 toolchain

patchset2: net: designware: fix descriptor layout and warnings on 64-bit archs

All members of the DMA descriptor must be strictly 32-bit, even on 64-bit
architectures: change the type to u32 to ensure this.

+++ cmd/pman_protection.c
@@ -181,7 +181,7 @@ static int pman_add_one_region(struct ptbl_hdr* tbl, struct ptab_rgn_body* rgn)
 
 static int setup_pman_by_armor(struct ptbl_hdr* tbl)
 {
-#define TO_PHY_ADDR(v) ((uint32_t)(v))
+#define TO_PHY_ADDR(v) ((uint32_t)(uintptr_t)(v)) /*use 32-bit for phyaddr*/
 	flush_cache((ulong)tbl, PTAB_LENGTH(tbl));
 	return secure_set_mem_protection(TO_PHY_ADDR(tbl), PTAB_LENGTH(tbl));
 #undef TO_PHY_ADDR
